requirements.txt
a2a-sdk==0.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
Authlib==1.6.0
bitarray==3.4.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.4
cytoolz==1.0.1
docstring_parser==0.16
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.115.14
frozenlist==1.7.0
google-adk==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.100.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.23.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-reflection==1.71.2
grpcio-status==1.71.2
grpcio-tools==1.71.2
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mcp==1.10.1
multidict==6.6.2
numpy==2.3.1
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
parsimonious==0.10.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
rpds-py==0.25.1
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
tenacity==8.5.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
uvicorn==0.35.0
web3==7.12.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
